+++ packages/Bio/src/package.ts
@@ -11,9 +11,7 @@ import {DimReductionBaseEditor, PreprocessFunctionReturnType}
 import {getActivityCliffs} from '@datagrok-libraries/ml/src/viewers/activity-cliffs';
 import {MmDistanceFunctionsNames} from '@datagrok-libraries/ml/src/macromolecule-distance-functions';
 import {BitArrayMetrics, KnownMetrics} from '@datagrok-libraries/ml/src/typed-metrics';
-import {
-  TAGS as bioTAGS,
-} from '@datagrok-libraries/bio/src/utils/macromolecule';
+import {NOTATION, TAGS as bioTAGS} from '@datagrok-libraries/bio/src/utils/macromolecule';
 import {SeqHandler, SeqTemps} from '@datagrok-libraries/bio/src/utils/seq-handler';
 import {IMonomerLib} from '@datagrok-libraries/bio/src/types';
 import {SeqPalette} from '@datagrok-libraries/bio/src/seq-palettes';
@@ -51,7 +49,7 @@ import {demoBio01bUI} from './demo/bio01b-hierarchical-clustering-and-activity-c
 import {demoBio03UI} from './demo/bio03-atomic-level';
 import {demoBio05UI} from './demo/bio05-helm-msa-sequence-space';
 import {checkInputColumnUI} from './utils/check-input-column';
-import {multipleSequenceAlignmentUI} from './utils/multiple-sequence-alignment-ui';
+import {MsaWarning, multipleSequenceAlignmentUI} from './utils/multiple-sequence-alignment-ui';
 import {WebLogoApp} from './apps/web-logo-app';
 import {SplitToMonomersFunctionEditor} from './function-edtiors/split-to-monomers-editor';
 import {splitToMonomersUI} from './utils/split-to-monomers';
@@ -75,6 +73,8 @@ import {DimReductionMethods} from '@datagrok-libraries/ml/src/multi-column-dimen
 import {
   ITSNEOptions, IUMAPOptions
 } from '@datagrok-libraries/ml/src/multi-column-dimensionality-reduction/multi-column-dim-reducer';
+import {generateLongSequence, generateLongSequence2} from '@datagrok-libraries/bio/src/utils/generator';
+
 import {CyclizedNotationProvider} from './utils/cyclized';
 import {getMolColumnFromHelm} from './utils/helm-to-molfile/utils';
 
@@ -598,7 +598,15 @@ export async function toAtomicLevel(table: DG.DataFrame, seqCol: DG.Column, nonl
 //description: Performs multiple sequence alignment
 //tags: bio, panel
 export function multipleSequenceAlignmentDialog(): void {
-  multipleSequenceAlignmentUI();
+  multipleSequenceAlignmentUI()
+    .catch((err: any) => {
+      const [errMsg, _errStack] = errInfo(err);
+      if (err instanceof MsaWarning) {
+        _package.logger.warning(errMsg);
+        return;
+      }
+      throw err;
+    });
 }
 
 //name: Multiple Sequence Alignment
@@ -944,6 +952,26 @@ export async function getRegionHelmApp(): Promise<void> {
   } finally {
     pi.close();
   }
+}
+
+// -- Tests long seq --
+
+//name: longSeqTableSeparator
+export function longSeqTableSeparator(): void {
+  const df = DG.DataFrame.fromColumns(generateLongSequence());
+  grok.shell.addTableView(df);
+}
+
+//name: longSeqTableFasta
+export function longSeqTableFasta(): void {
+  const df = DG.DataFrame.fromColumns([generateLongSequence2(NOTATION.FASTA)]);
+  grok.shell.addTableView(df);
+}
+
+//name: longSeqTableHelm
+export function longSeqTableHelm(): void {
+  const df = DG.DataFrame.fromColumns([generateLongSequence2(NOTATION.HELM)]);
+  grok.shell.addTableView(df);
 }
 
 // -- Handle context menu --