+++ packages/Bio/src/tests/substructure-filters-tests.ts
@@ -186,7 +186,7 @@ category('bio-substructure-filters', async () => {
     }
     await filter.awaitRendered();
     await delay(3000); //TODO: await for grid.onLookChanged
-  }, {skipReason: 'GROK-15678'});
+  });
 
   // Generates unhandled exception accessing isFiltering before bioFilter created
   test('helm-view', async () => {
@@ -310,7 +310,7 @@ category('bio-substructure-filters', async () => {
     await Promise.all([f1.awaitRendered(), f2.awaitRendered()]);
     await awaitGrid(view.grid);
     await delay(3000); //TODO: await for grid.onLookChanged
-  }, {skipReason: 'GROK-15678'});
+  });
 
   // two seq columns